refactor: move asyncHandler from middlewares to utils
The asyncHandler utility function is more appropriately placed in the utils directory since it's a general utility rather than middleware-specific functionality. This change improves code organization and maintainability.
